requirements.txt
openai
python-dotenv
python-telegram-bot